[py math] Start to split bindings into separate files
This loses the easy-to-find git history for matmul, but it's fairly
new (so not too much is lost) and tracing it across the file shuffle
isn't too difficult, anyway.
