+++ bindings/pydrake/math/math_py.h
@@ -15,8 +15,17 @@ namespace drake {
 namespace pydrake {
 namespace internal {
 
+/* Binds a native C++ matmul(A, B) for wide variety of scalar types. This is
+important for performance until numpy allows us to define dtype-specific
+implementations. Doing a matmul elementwise with a C++ <=> Python call for every
+flop is extraordinarily slow. */
+void DefineMathMatmul(py::module m);
+
 /* Defines bindings per math_py_monolith.cc. */
 void DefineMathMonolith(py::module m);
+
+/* Defines bindings per math_py_operators.cc. */
+void DefineMathOperators(py::module m);
 
 }  // namespace internal
 }  // namespace pydrake
